Move ReadOnlyContext to Pickler
