+++ compiler/src/dotty/tools/io/FileWriters.scala
@@ -35,118 +35,6 @@ object FileWriters {
   private def classRelativePath(className: String, suffix: String): String =
     className.replace('.', '/') + suffix
 
-  inline def ctx(using ReadOnlyContext): ReadOnlyContext = summon[ReadOnlyContext]
-
-  sealed trait DelayedReporter {
-    def hasErrors: Boolean
-    def error(message: Context ?=> Message, position: SourcePosition): Unit
-    def warning(message: Context ?=> Message, position: SourcePosition): Unit
-    def log(message: String): Unit
-
-    final def toBuffered: Option[BufferingReporter] = this match
-      case buffered: BufferingReporter =>
-        if buffered.hasReports then Some(buffered) else None
-      case _: EagerReporter => None
-
-    def error(message: Context ?=> Message): Unit = error(message, NoSourcePosition)
-    def warning(message: Context ?=> Message): Unit = warning(message, NoSourcePosition)
-    final def exception(reason: Context ?=> Message, throwable: Throwable): Unit =
-      error({
-        val trace = throwable.getStackTrace().mkString("\n  ")
-        em"An unhandled exception was thrown in the compiler while\n  ${reason.message}.\n${throwable}\n  $trace"
-      }, NoSourcePosition)
-  }
-
-  final class EagerReporter(using captured: Context) extends DelayedReporter:
-    private var _hasErrors = false
-
-    def hasErrors: Boolean = _hasErrors
-
-    def error(message: Context ?=> Message, position: SourcePosition): Unit =
-      report.error(message, position)
-      _hasErrors = true
-
-    def warning(message: Context ?=> Message, position: SourcePosition): Unit =
-      report.warning(message, position)
-
-    def log(message: String): Unit = report.echo(message)
-
-  enum Report:
-    case Error(message: Context => Message, position: SourcePosition)
-    case Warning(message: Context => Message, position: SourcePosition)
-    case Log(message: String)
-
-  final class BufferingReporter extends DelayedReporter {
-    // We optimise access to the buffered reports for the common case - that there are no warning/errors to report
-    // We could use a listBuffer etc - but that would be extra allocation in the common case
-    // buffered logs are updated atomically.
-
-    private val _bufferedReports = AtomicReference(List.empty[Report])
-    private val _hasErrors = AtomicBoolean(false)
-
-
-    /** Atomically record that an error occurred */
-    private def recordError(): Unit =
-      _hasErrors.set(true)
-
-    /** Atomically add a report to the log */
-    private def recordReport(report: Report): Unit =
-      _bufferedReports.getAndUpdate(report :: _)
-
-    /** atomically extract and clear the buffered reports, must only be called at a synchronization point. */
-    def resetReports(): List[Report] =
-      val curr = _bufferedReports.get()
-      if curr.nonEmpty && !_bufferedReports.compareAndSet(curr, Nil) then
-        throw ConcurrentModificationException("concurrent modification of buffered reports")
-      else curr
-
-    def hasErrors: Boolean = _hasErrors.get()
-    def hasReports: Boolean = _bufferedReports.get().nonEmpty
-
-    def error(message: Context ?=> Message, position: SourcePosition): Unit =
-      recordReport(Report.Error({case given Context => message}, position))
-      recordError()
-
-    def warning(message: Context ?=> Message, position: SourcePosition): Unit =
-      recordReport(Report.Warning({case given Context => message}, position))
-
-    def log(message: String): Unit =
-      recordReport(Report.Log(message))
-  }
-
-  trait ReadOnlySettings:
-    def jarCompressionLevel: Int
-    def debug: Boolean
-
-  trait ReadOnlyRun:
-    def suspendedAtTyperPhase: Boolean
-
-  trait ReadOnlyContext:
-    val run: ReadOnlyRun
-    val settings: ReadOnlySettings
-    val reporter: DelayedReporter
-
-  trait BufferedReadOnlyContext extends ReadOnlyContext:
-    val reporter: BufferingReporter
-
-  object ReadOnlyContext:
-    def readSettings(using ctx: Context): ReadOnlySettings = new:
-      val jarCompressionLevel = ctx.settings.XjarCompressionLevel.value
-      val debug = ctx.settings.Ydebug.value
-
-    def readRun(using ctx: Context): ReadOnlyRun = new:
-      val suspendedAtTyperPhase = ctx.run.nn.suspendedAtTyperPhase
-
-    def buffered(using Context): BufferedReadOnlyContext = new:
-      val settings = readSettings
-      val reporter = BufferingReporter()
-      val run = readRun
-
-    def eager(using Context): ReadOnlyContext = new:
-      val settings = readSettings
-      val reporter = EagerReporter()
-      val run = readRun
-
   /**
    * The interface to writing classfiles. GeneratedClassHandler calls these methods to generate the
    * directory and files that are created, and eventually calls `close` when the writing is complete.
@@ -162,7 +50,7 @@ object FileWriters {
      *
      * @param name the internal name of the class, e.g. "scala.Option"
      */
-    def writeTasty(name: String, bytes: Array[Byte])(using ReadOnlyContext): AbstractFile
+    def writeTasty(name: String, bytes: Array[Byte])(using Context): AbstractFile
 
     /**
      * Close the writer. Behavior is undefined after a call to `close`.
@@ -172,14 +60,14 @@ object FileWriters {
 
   object TastyWriter {
 
-    def apply(output: AbstractFile)(using ReadOnlyContext): TastyWriter =
+    def apply(output: AbstractFile)(using Context): TastyWriter =
       // In Scala 2 depending on cardinality of distinct output dirs MultiClassWriter could have been used
       // In Dotty we always use single output directory
       new SingleTastyWriter(FileWriter(output, None))
 
     private final class SingleTastyWriter(underlying: FileWriter) extends TastyWriter {
 
-      override def writeTasty(className: String, bytes: Array[Byte])(using ReadOnlyContext): AbstractFile = {
+      override def writeTasty(className: String, bytes: Array[Byte])(using Context): AbstractFile = {
         underlying.writeFile(classRelativePath(className, ".tasty"), bytes)
       }
 
@@ -202,7 +90,7 @@ object FileWriters {
     /**
      * Write a classfile
      */
-    def writeClass(name: String, bytes: Array[Byte])(using ReadOnlyContext): AbstractFile
+    def writeClass(name: String, bytes: Array[Byte])(using Context): AbstractFile
 
     /**
      * Close the writer. Behavior is undefined after a call to `close`.
@@ -211,7 +99,7 @@ object FileWriters {
   }
 
   object ClassfileWriter {
-    def apply(output: AbstractFile, jarManifestMainClass: Option[String], dumpClassesPath: Option[AbstractFile])(using ReadOnlyContext): ClassfileWriter = {
+    def apply(output: AbstractFile, jarManifestMainClass: Option[String], dumpClassesPath: Option[AbstractFile])(using Context): ClassfileWriter = {
       // In Scala 2 depending on cardinality of distinct output dirs MultiClassWriter could have been used
       // In Dotty we always use single output directory
       val basicClassWriter = new SingleClassWriter(FileWriter(output, jarManifestMainClass))
@@ -221,11 +109,11 @@ object FileWriters {
     }
 
     private final class SingleClassWriter(underlying: FileWriter) extends ClassfileWriter {
-      override def writeClass(className: String, bytes: Array[Byte])(using ReadOnlyContext): AbstractFile = {
+      override def writeClass(className: String, bytes: Array[Byte])(using Context): AbstractFile = {
         underlying.writeFile(classRelativePath(className, ".class"), bytes)
       }
 
-      override def writeTasty(className: String, bytes: Array[Byte])(using ReadOnlyContext): AbstractFile = {
+      override def writeTasty(className: String, bytes: Array[Byte])(using Context): AbstractFile = {
         underlying.writeFile(classRelativePath(className, ".tasty"), bytes)
       }
 
@@ -233,13 +121,13 @@ object FileWriters {
     }
 
     private final class DebugClassWriter(basic: ClassfileWriter, dump: FileWriter) extends ClassfileWriter {
-      override def writeClass(className: String, bytes: Array[Byte])(using ReadOnlyContext): AbstractFile = {
+      override def writeClass(className: String, bytes: Array[Byte])(using Context): AbstractFile = {
         val outFile = basic.writeClass(className, bytes)
         dump.writeFile(classRelativePath(className, ".class"), bytes)
         outFile
       }
 
-      override def writeTasty(className: String, bytes: Array[Byte])(using ReadOnlyContext): AbstractFile = {
+      override def writeTasty(className: String, bytes: Array[Byte])(using Context): AbstractFile = {
         basic.writeTasty(className, bytes)
       }
 
@@ -252,12 +140,12 @@ object FileWriters {
 
 
   sealed trait FileWriter {
-    def writeFile(relativePath: String, bytes: Array[Byte])(using ReadOnlyContext): AbstractFile
+    def writeFile(relativePath: String, bytes: Array[Byte])(using Context): AbstractFile
     def close(): Unit
   }
 
   object FileWriter {
-    def apply(file: AbstractFile, jarManifestMainClass: Option[String])(using ReadOnlyContext): FileWriter =
+    def apply(file: AbstractFile, jarManifestMainClass: Option[String])(using Context): FileWriter =
       if (file.isInstanceOf[JarArchive]) {
         val jarCompressionLevel = ctx.settings.jarCompressionLevel
         // Writing to non-empty JAR might be an undefined behaviour, e.g. in case if other files where
@@ -296,7 +184,7 @@ object FileWriters {
 
     lazy val crc = new CRC32
 
-    override def writeFile(relativePath: String, bytes: Array[Byte])(using ReadOnlyContext): AbstractFile = this.synchronized {
+    override def writeFile(relativePath: String, bytes: Array[Byte])(using Context): AbstractFile = this.synchronized {
       val entry = new ZipEntry(relativePath)
       if (storeOnly) {
         // When using compression method `STORED`, the ZIP spec requires the CRC and compressed/
@@ -330,14 +218,14 @@ object FileWriters {
     val noAttributes = Array.empty[FileAttribute[?]]
     private val isWindows = scala.util.Properties.isWin
 
-    private def checkName(component: Path)(using ReadOnlyContext): Unit = if (isWindows) {
+    private def checkName(component: Path)(using Context): Unit = if (isWindows) {
       val specials = raw"(?i)CON|PRN|AUX|NUL|COM[1-9]|LPT[1-9]".r
       val name = component.toString
       def warnSpecial(): Unit = ctx.reporter.warning(em"path component is special Windows device: ${name}")
       specials.findPrefixOf(name).foreach(prefix => if (prefix.length == name.length || name(prefix.length) == '.') warnSpecial())
     }
 
-    def ensureDirForPath(baseDir: Path, filePath: Path)(using ReadOnlyContext): Unit = {
+    def ensureDirForPath(baseDir: Path, filePath: Path)(using Context): Unit = {
       import java.lang.Boolean.TRUE
       val parent = filePath.getParent
       if (!builtPaths.containsKey(parent)) {
@@ -367,7 +255,7 @@ object FileWriters {
     private val fastOpenOptions = util.EnumSet.of(StandardOpenOption.CREATE_NEW, StandardOpenOption.WRITE)
     private val fallbackOpenOptions = util.EnumSet.of(StandardOpenOption.CREATE, StandardOpenOption.WRITE, StandardOpenOption.TRUNCATE_EXISTING)
 
-    override def writeFile(relativePath: String, bytes: Array[Byte])(using ReadOnlyContext): AbstractFile = {
+    override def writeFile(relativePath: String, bytes: Array[Byte])(using Context): AbstractFile = {
       val path = base.resolve(relativePath)
       try {
         ensureDirForPath(base, path)
@@ -417,7 +305,7 @@ object FileWriters {
       finally out.close()
     }
 
-    override def writeFile(relativePath: String, bytes: Array[Byte])(using ReadOnlyContext): AbstractFile = {
+    override def writeFile(relativePath: String, bytes: Array[Byte])(using Context): AbstractFile = {
       val outFile = getFile(base, relativePath)
       writeBytes(outFile, bytes)
       outFile
